refactor: ExperienceSection 레이아웃 변경 컴포넌트화
