+++ gitlio/app/editor/_components/mainSection/experienceLayoutOptions/ExperienceOptionOne.tsx
@@ -1,0 +1,46 @@
+'use client';
+import React from 'react';
+import { MdBrightness1 } from 'react-icons/md';
+import cx from 'classnames';
+import { Section } from '@/store/experienceSectionStore';
+
+interface ExperienceOptionOneProps {
+  sections: Section[];
+}
+
+const ExperienceOptionOne: React.FC<ExperienceOptionOneProps> = ({
+  sections,
+}) => {
+  return (
+    <ul className="timeline timeline-snap-icon timeline-compact timeline-vertical px-10 py-10">
+      {sections.map((item, index) => (
+        <li key={item.id}>
+          {index !== 0 && <hr />}
+          <div className="timeline-middle">
+            <MdBrightness1 className="h-5 w-5" />
+          </div>
+          <div
+            className="timeline-end md:text-start mb-10"
+            style={{ whiteSpace: 'pre-wrap' }}
+          >
+            <time className="font-mono italic">
+              {item.startDate} - {item.endDate}
+            </time>
+            <div className="text-lg font-black">{item.title}</div>
+            <div
+              className={cx(
+                item.description &&
+                  'mt-3 ml-3 bg-neutral-content/30 px-5 py-5 rounded-xl'
+              )}
+            >
+              {item.description}
+            </div>
+          </div>
+          <hr />
+        </li>
+      ))}
+    </ul>
+  );
+};
+
+export default ExperienceOptionOne;
