+++ gitlio/app/editor/_components/mainSection/introLayoutOptions/IntroOptionOne.tsx
