+++ gitlio/app/editor/_components/mainSection/introLayoutOptions/IntroOptionTwo.tsx
